Move MaxMind downloader under MaxMind namespace
