+++ .travis.yml
@@ -17,7 +17,7 @@ before_install:
 
 script:
   - RAILS_ENV=test bin/rake db:drop db:create db:schema:load
-  - bin/rake download_max_mind_db
+  - bin/rake max_mind:download
   - bin/rake
 
 before_script:
